requirements.txt
Django>=1.8.3
django-parler==1.5
marshmallow==2.0.0b4
purl==1.1
    six==1.9.0
wheel==0.24.0
